* HTTP API accessor can now list all devices and groups
* devices now also contain information about temperature, power, etc.
* fixed some mistakes in the data types of the HTTP API accessor

+++ FritzBoxAPI/FritzBoxHttpAPI/v1.09/TemperatureDTO.cs
@@ -10,9 +10,10 @@ namespace XElement.RedYellowBlue.FritzBoxAPI.FritzBoxHttpAPI.v109
 
         /// <summary>
         /// celsius: Wert in 0,1 °C, negative und positive Werte möglich
+        /// Actual example: "220" for "22,0°C"
         /// </summary>
         [XmlElement( "celsius" )]
-        public double Celsius { get; set; }
+        public int Celsius { get; set; }
 
 
         /// <summary>
